Move reading statistics out of DeltaLakeMetastore
This removes `DeltaLakeMetastore.getTableStatistics` method, moving the
logic to a new class.
