+++ plugin/trino-delta-lake/src/main/java/io/trino/plugin/deltalake/DeltaLakeMetadataFactory.java
@@ -17,6 +17,7 @@ import io.airlift.json.JsonCodec;
 import io.trino.filesystem.TrinoFileSystemFactory;
 import io.trino.plugin.deltalake.metastore.HiveMetastoreBackedDeltaLakeMetastore;
 import io.trino.plugin.deltalake.statistics.CachingExtendedStatisticsAccess;
+import io.trino.plugin.deltalake.statistics.FileBasedTableStatisticsProvider;
 import io.trino.plugin.deltalake.transactionlog.TransactionLogAccess;
 import io.trino.plugin.deltalake.transactionlog.checkpoint.CheckpointWriterManager;
 import io.trino.plugin.deltalake.transactionlog.writer.TransactionLogWriterFactory;
@@ -110,9 +111,12 @@ public class DeltaLakeMetadataFactory
         HiveMetastoreBackedDeltaLakeMetastore deltaLakeMetastore = new HiveMetastoreBackedDeltaLakeMetastore(
                 cachingHiveMetastore,
                 transactionLogAccess,
-                typeManager,
                 statisticsAccess,
                 fileSystemFactory);
+        FileBasedTableStatisticsProvider tableStatisticsProvider = new FileBasedTableStatisticsProvider(
+                typeManager,
+                transactionLogAccess,
+                statisticsAccess);
         TrinoViewHiveMetastore trinoViewHiveMetastore = new TrinoViewHiveMetastore(
                 cachingHiveMetastore,
                 accessControlMetadata.isUsingSystemSecurity(),
@@ -121,6 +125,7 @@ public class DeltaLakeMetadataFactory
         return new DeltaLakeMetadata(
                 deltaLakeMetastore,
                 transactionLogAccess,
+                tableStatisticsProvider,
                 fileSystemFactory,
                 typeManager,
                 accessControlMetadata,
